+++ frontend/src/pages/Auth/RegisterPage/index.tsx
@@ -2,14 +2,14 @@ import { FC, useEffect, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import "./LoginPage.css";
 //importing svg
-import Logo from "../components/Logo";
-import Google from "../components/Google";
-import Github from "../components/Github";
+import Logo from "../../../components/Logo";
+import Google from "../../../components/Google";
+import Github from "../../../components/Github";
 
 //importing icons
 import { BiUser } from "react-icons/bi";
 import { AiOutlineUnlock } from "react-icons/ai";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../../../context/AuthContext";
 
 export const RegisterPage: FC = () => {
   const [email, setEmail] = useState("");
